moved apierror classes into package, removed @notnull in user.password
