+++ src/main/java/com/travelstory/entity/User.java
@@ -7,7 +7,6 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import javax.persistence.*;
 import javax.validation.constraints.Email;
-import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -28,7 +27,6 @@ public class User {
     @Email
     private String email;
 
-    @NotNull
     @Size(min = 1, max = 25)
     private String password;
