--- src/main/java/com/travelstory/exceptions/ApiSubError.java
@@ -1,5 +1,0 @@
-package com.travelstory.exceptions;
-
-abstract class ApiSubError {
-
-}
