+++ src/main/java/com/travelstory/exceptions/api/ApiSubError.java
@@ -1,0 +1,5 @@
+package com.travelstory.exceptions.api;
+
+abstract class ApiSubError {
+
+}
