+++ src/main/java/com/travelstory/exceptions/api/ApiValidationError.java
@@ -1,4 +1,4 @@
-package com.travelstory.exceptions;
+package com.travelstory.exceptions.api;
 
 import java.util.Objects;
